build: Move vcpkg buildtrees/install roots to target/ and add a MAX_PATH check
In order to keep the src/ tree as clean as possible, move vcpkg buildtrees and
install roots to be under the build output under target/. This helps prevent
caching issues when vcpkg srcs and dependencies change.

In addition, add a max path length check to buildtrees root path, as it can
contain very long paths that break CMake builds.